solution/requirements.txt
numpy
pandas
sentence-transformers
faiss-cpu
tqdm
rank-bm25
